+++ build-system/tasks/presubmit-checks.js
@@ -176,7 +176,7 @@ const forbiddenTerms = {
   'installActionServiceForDoc': {
     message: privateServiceFactory,
     whitelist: [
-      'src/inabox/amp-inabox.js',
+      'src/inabox/inabox-services.js',
       'src/service/action-impl.js',
       'src/service/core-services.js',
       'src/service/standard-actions-impl.js',
@@ -257,7 +257,7 @@ const forbiddenTerms = {
   'installUrlReplacementsServiceForDoc': {
     message: privateServiceFactory,
     whitelist: [
-      'src/inabox/amp-inabox.js',
+      'src/inabox/inabox-services.js',
       'src/service/core-services.js',
       'src/service/url-replacements-impl.js',
     ],
@@ -266,7 +266,7 @@ const forbiddenTerms = {
     message: privateServiceFactory,
     whitelist: [
       'src/runtime.js',
-      'src/inabox/amp-inabox.js',
+      'src/inabox/inabox-services.js',
       'src/service/core-services.js',
       'src/service/viewer-impl.js',
     ],
@@ -300,7 +300,7 @@ const forbiddenTerms = {
   'installResourcesServiceForDoc': {
     message: privateServiceFactory,
     whitelist: [
-      'src/inabox/amp-inabox.js',
+      'src/inabox/inabox-services.js',
       'src/service/core-services.js',
       'src/service/resources-impl.js',
       'src/service/standard-actions-impl.js',
